+++ src/app-store/actions.js
@@ -1,6 +1,43 @@
-import {getByArraySyntax} from "../../src/utils/fetchFirstModuleConfig";
+import {getByArraySyntax} from "../utils/fetchFirstModuleConfig";
 
 export default {
+    /**
+     * Copies the the content of the given element to the clipboard if the browser accepts the command.
+     * Solution for the weird behaviour on iOS from:
+     * https://stackoverflow.com/questions/34045777/copy-to-clipboard-using-javascript-in-ios
+     *
+     * @param {Element} el element to copy,
+     * @returns {void}
+     */
+    copyToClipboard ({dispatch}, el) {
+        const oldReadOnly = el.readOnly,
+            oldContentEditable = el.contentEditable,
+            range = document.createRange(),
+            selection = window.getSelection();
+
+        el.readOnly = false;
+        el.contentEditable = true;
+
+        range.selectNodeContents(el);
+        selection.removeAllRanges();
+        if (!Radio.request("Util", "isInternetExplorer")) {
+            selection.addRange(range);
+        }
+        // Seems to be required for mobile devices
+        el.setSelectionRange(0, 999999);
+
+        el.readOnly = oldReadOnly;
+        el.contentEditable = oldContentEditable;
+
+        try {
+            document.execCommand("copy");
+            dispatch("Alerting/addSingleAlert", {content: i18next.t("common:modules.util.copyToClipboard.contentSaved")}, {root: true});
+        }
+        catch (err) {
+            dispatch("Alerting/addSingleAlert", {content: i18next.t("common:modules.util.copyToClipboard.contentNotSaved")}, {root: true});
+            console.error(`CopyToClipboard: ${err}`);
+        }
+    },
     /**
      * Function to check if the deprecated parameters could be specified for more than one location e.g. they (location of the parameter or tool) have multiple possible paths.
     * Furthermore the function checks whether the given paths for the parameters are defined or undefined.