+++ src/modules/tools/indexTools.js
@@ -10,6 +10,8 @@ import actions from "./actionsTools";
 import Draw from "./draw/store/indexDraw";
 import FileImport from "./fileImport/store/indexFileImport";
 import Gfi from "./gfi/store/indexGfi";
+import Measure from "./measure/store/indexMeasure";
+import SaveSelection from "./saveSelection/store/indexSaveSelection";
 import ScaleSwitcher from "./scaleSwitcher/store/indexScaleSwitcher";
 import StyleVT from "./styleVT/store/indexStyleVT";
 import SupplyCoord from "./supplyCoord/store/indexSupplyCoord";
@@ -24,6 +26,8 @@ export default {
         Draw,
         FileImport,
         Gfi,
+        Measure,
+        SaveSelection,
         ScaleSwitcher,
         StyleVT,
         SupplyCoord