+++ src/modules/tools/stateTools.js
@@ -1,6 +1,8 @@
 import Draw from "./draw/components/Draw.vue";
 import FileImport from "./fileImport/components/FileImport.vue";
 import Gfi from "./gfi/components/Gfi.vue";
+import Measure from "./measure/components/Measure.vue";
+import SaveSelection from "./saveSelection/components/SaveSelection.vue";
 import ScaleSwitcher from "./scaleSwitcher/components/ScaleSwitcher.vue";
 import StyleVT from "./styleVT/components/StyleVT.vue";
 import SupplyCoord from "./supplyCoord/components/SupplyCoord.vue";
@@ -15,6 +17,8 @@ const state = {
         draw: Draw,
         fileImport: FileImport,
         gfi: Gfi,
+        measure: Measure,
+        saveSelection: SaveSelection,
         scaleSwitcher: ScaleSwitcher,
         styleVT: StyleVT,
         supplyCoord: SupplyCoord